A2A/Simple_A2A/requirements.txt
uvicorn 
fastapi 
httpx 
pydantic 
a2a-sdk[http-server]>=0.3.0
langgraph 
google-adk
python-dotenv